Move entities
