+++ app/src/main/java/com/example/emmaburton/projectemma/data/api/RecipesApi.kt
@@ -1,6 +1,6 @@
 package com.example.emmaburton.projectemma.data.api
 
-import com.example.emmaburton.projectemma.data.model.Recipe
+import com.example.emmaburton.projectemma.entities.Recipe
 import com.google.gson.GsonBuilder
 import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
